+++ Classes/SiteManager.h
@@ -15,29 +15,29 @@ class PorkerManager;
 class SiteManager
 {
 public:
-	SiteManager(Node* parent,unsigned long long currentPlayerID);
+	SiteManager(Node* parent,int currentPlayerID);
 	virtual ~SiteManager();
 
 	//玩家找个位置坐下
-	bool joinSite(unsigned long long playerID, string playerName, int diamond, int money);
+	bool joinSite(int playerID, string playerName, int diamond, int money);
 	//离开座位
-	bool leaveSite(unsigned long long playerID);
+	bool leaveSite(int playerID);
 	//设置倍数
-	bool setMultiple(unsigned long long playerID, int multiple);
+	bool setMultiple(int playerID, int multiple);
 	//显示倍数
-	bool showMultiple(unsigned long long playerID,bool isShow = true);
+	bool showMultiple(int playerID, bool isShow = true);
 	//设置庄家
-	bool showZhuangJia(unsigned long long playerID);
+	bool showZhuangJia(int playerID);
 	//显示牛
-	bool showNiu(unsigned long long playerID, int niuIndex);
+	bool showNiu(int playerID, int niuIndex);
 	//隐藏所有牛
 	void hideAllNiu();
 	//显示准备
-	bool showReady(unsigned long long playerID, bool isShow = true);
+	bool showReady(int playerID, bool isShow = true);
 	//当前入座玩家数量
 	int currentPlayerCount() const;
 	//显示聊天信息
-	void showChatMessage(unsigned long long playerID, string strMessage, float timeThreshold = 3.0f);
+	void showChatMessage(int playerID, string strMessage, float timeThreshold = 3.0f);
 
 private:
 	//玩家和头像位置信息
@@ -55,11 +55,11 @@ private:
 	array<PlayerAndProfilePos,5> m_playerProfileInfo;			//位置信息
 	array<IUserProfileUIInGame*,5> m_pUserProfileVecs;			//头像UI，第0个是当前玩家头像
 	array<int,5> m_playerInRoom;								//房间里的坑位，0表示没人坐，1表示有人坐
-	array<unsigned long long,5> m_inRoomPlayerID;				//0为不存在
+	array<int, 5> m_inRoomPlayerID;				//0为不存在
 	array<Sprite*,5> m_playerNiu;
 	map<int, string> m_niuIndex2JPGPath;
 
-	unsigned long long m_currentPlayerID;						//自己的ID
+	int m_currentPlayerID;						//自己的ID
 
 	Node* m_pParent;
